boj: solve 24267 알고리즘 수업 - 알고리즘의 수행 시간 6
- https://www.acmicpc.net/problem/24267

+++ src/boj/p24k/p24264.rs
@@ -3,7 +3,7 @@ use std::io::{BufRead, Write};
 
 #[allow(dead_code)]
 fn solve24264(reader: &mut impl BufRead, writer: &mut impl Write) {
-    let n = read_line(reader).parse::<usize>().unwrap();
+    let n = read_line(reader).parse::<i64>().unwrap();
 
     writeln!(writer, "{}", n * n).unwrap();
     writeln!(writer, "2").unwrap();
