+++ src/boj/p25k/p25314.rs
@@ -3,7 +3,7 @@ use std::io::{BufRead, Write};
 
 #[allow(dead_code)]
 fn solve25314(reader: &mut impl BufRead, writer: &mut impl Write) {
-    let n = read_line(reader).parse::<usize>().unwrap();
+    let n = read_line(reader).parse::<i64>().unwrap();
 
     for _ in 0..(n / 4) {
         write!(writer, "long ").unwrap();
